Move asset certification code into a new separate crate
This PR moves the asset certification code to a new crate,
later to be reused by all asset serving canisters in this repository.

This PR intentionally only _moves_ code (with the only exception of
switching out the implementation for the representation independent
hash). There will be a subsequent PR to make the library interface nicer.

The code was only moved from `assets.rs` with the exception of the 
`ContentType` impl. The functionality for certification headers that
 currently resides in `http.rs` will be added to the crate in later PR as well.
