+++ Cargo.toml
@@ -1,5 +1,6 @@
 [workspace]
 members = [
+    "src/asset_util",
     "src/internet_identity",
     "src/canister_sig_util",
     "src/canister_tests",
@@ -17,6 +18,7 @@ opt-level = 'z'
 
 [workspace.dependencies]
 # local dependencies
+asset_util = { path = "src/asset_util" }
 canister_sig_util = { path = "src/canister_sig_util" }
 canister_tests = { path = "src/canister_tests" }
 internet_identity_interface = { path = "src/internet_identity_interface" }
@@ -28,6 +30,7 @@ ic-cdk = "0.10"
 ic-cdk-macros = "0.7"
 ic-certification = "1.3"
 ic-metrics-encoder = "1"
+ic-representation-independent-hash = "1.3"
 ic-response-verification = "1.3"
 ic-stable-structures = "0.5"
 ic-test-state-machine-client = "3"
