+++ Dockerfile
@@ -49,6 +49,7 @@ COPY src/archive/Cargo.toml src/archive/Cargo.toml
 COPY src/canister_tests/Cargo.toml src/canister_tests/Cargo.toml
 COPY src/canister_sig_util/Cargo.toml src/canister_sig_util/Cargo.toml
 COPY src/vc_util/Cargo.toml src/vc_util/Cargo.toml
+COPY src/asset_util/Cargo.toml src/asset_util/Cargo.toml
 ENV CARGO_TARGET_DIR=/cargo_target
 COPY ./scripts/build ./scripts/build
 RUN mkdir -p src/internet_identity/src \
@@ -63,6 +64,8 @@ RUN mkdir -p src/internet_identity/src \
     && touch src/canister_sig_util/src/lib.rs \
     && mkdir -p src/vc_util/src \
     && touch src/vc_util/src/lib.rs \
+    && mkdir -p src/asset_util/src \
+    && touch src/asset_util/src/lib.rs \
     && ./scripts/build --only-dependencies --internet-identity --archive \
     && rm -rf src
