+++ src/asset_util/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "asset_util"
+description = "Asset certification utility library for asset serving canisters"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+base64.workspace = true
+candid.workspace = true
+ic-cdk.workspace = true
+ic-cdk-macros.workspace = true
+ic-certification.workspace = true
+ic-representation-independent-hash.workspace = true
+include_dir = "0.7"
+internet_identity_interface.workspace = true
+lazy_static.workspace = true
+sha2.workspace = true
