+++ src/internet_identity/src/vc_mvp.rs
@@ -1,4 +1,3 @@
-use crate::assets::CertifiedAssets;
 use crate::delegation::check_frontend_length;
 use crate::{delegation, hash, state, update_root_hash, LABEL_SIG, MINUTE_NS};
 use candid::Principal;
@@ -10,6 +9,7 @@ use identity_core::common::{Timestamp, Url};
 use identity_core::convert::FromJson;
 use identity_credential::credential::{Credential, CredentialBuilder, Subject};
 
+use asset_util::CertifiedAssets;
 use canister_sig_util::signature_map::SignatureMap;
 use internet_identity_interface::internet_identity::types::vc_mvp::{
     GetIdAliasResponse, IdAliasCredentials, PreparedIdAlias, SignedIdAlias,
